Support deploying bash in Cloud Foundry via GoTTY
Added a manifest file that is used by default by Cloud Foundry's cli
`cf push` command. This manifest causes GoTTY to launch bash and allow
it to accept keystrokes. We do not use GoTTY's -c parameter to force use
of credentials, because that requires that the user update the manifest
file before `cf push`, and also have the utilities to generate secure
password on her machine. We use the alternative method to obscure the
access to this app by recommending (in README) to use the --random-route
option.

Updated code to listen on environment variable PORT. This overrides
logic that purportedly supports choosing a different port via environment
variable GOTTY_PORT. But since that doesn't really work (see issue #122)
it's okay for us to hack this to support Cloud Foundry.